Wrapper for correct types
